requirements.txt
asgiref==3.3.1
boto3==1.17.53
botocore==1.20.53
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.1.5
django-cleanup==5.1.0
django-crispy-forms==1.10.0
django-embed-video==1.4.0
django-storages==1.11.1
idna==2.10
jmespath==0.10.0
oauthlib==3.1.0
Pillow==8.1.0
pycparser==2.20
pycrypto==2.6.1
PyJWT==2.0.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.7
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
twilio==6.51.1
urllib3==1.26.3
